+++ opensoundscape/__init__.py
@@ -5,7 +5,7 @@ from . import audio_tools
 from . import audio
 from . import aru
 from . import data_selection
-from . import helpers
+from . import utils
 from . import localization
 from . import metrics
 from . import ribbit
