+++ tests/test_annotations.py
@@ -5,7 +5,7 @@ from opensoundscape.annotations import BoxedAnnotations
 from pathlib import Path
 import pandas as pd
 import numpy as np
-from opensoundscape.helpers import generate_clip_times_df
+from opensoundscape.utils import generate_clip_times_df
 
 
 @pytest.fixture()
